Add history API support
