+++ src/comment.js
@@ -27,11 +27,20 @@ export default class Comment extends Component {
   async init() {
     const { depth, setUser, id } = this.props;
     const comment = await loadItem(id);
-    this.setState({
-      comment,
-      isDisplayed: comment && comment.type === 'comment' && comment.text,
-      kids: comment.kids.map((kId) => <Comment setUser={setUser} depth={depth + 1} key={kId} id={kId} />),
-    });
+    if (this._isMounted === true) {
+      this.setState({
+        comment,
+        isDisplayed: comment && comment.type === 'comment' && comment.text,
+        kids: comment.kids.map((kId) => <Comment setUser={setUser} depth={depth + 1} key={kId} id={kId} />),
+      });
+    }
+  }
+  componentDidMount() {
+    this._isMounted = true;
+  }
+
+  componentWillUnmount() {
+    this._isMounted = false;
   }
 
   toggle() {
